clean & refactoring
